Put components into their own individual files
